requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.1.0
Babel==2.9.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.14
django-cors-headers==4.4.0
django-gunicorn==0.1.1
django-jazzmin==3.0.0
djangorestframework==3.15.2
drf-yasg==1.21.7
environs==11.0.0
frozenlist==1.4.1
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
magic-filter==1.0.12
marshmallow==3.21.3
multidict==6.0.5
packaging==24.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
yarl==1.9.4
